Move some commands from `common` to the handler
