--- internal/handler/common/saslstart_plain.go
@@ -1,71 +1,0 @@
-// Copyright 2021 FerretDB Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package common
-
-import (
-	"bytes"
-	"encoding/base64"
-	"fmt"
-
-	"github.com/FerretDB/FerretDB/internal/handler/commonerrors"
-	"github.com/FerretDB/FerretDB/internal/types"
-)
-
-// saslStartPlain extracts username and password from PLAIN `saslStart` payload.
-func saslStartPlain(doc *types.Document) (string, string, error) {
-	var payload []byte
-
-	// some drivers send payload as a string
-	stringPayload, err := GetRequiredParam[string](doc, "payload")
-	if err == nil {
-		if payload, err = base64.StdEncoding.DecodeString(stringPayload); err != nil {
-			return "", "", commonerrors.NewCommandErrorMsgWithArgument(
-				commonerrors.ErrBadValue,
-				fmt.Sprintf("Invalid payload: %v", err),
-				"payload",
-			)
-		}
-	}
-
-	// most drivers follow spec and send payload as a binary
-	binaryPayload, err := GetRequiredParam[types.Binary](doc, "payload")
-	if err == nil {
-		payload = binaryPayload.B
-	}
-
-	// as spec's payload should be binary, we return an error mentioned binary as expected type
-	if payload == nil {
-		return "", "", err
-	}
-
-	parts := bytes.Split(payload, []byte{0})
-	if l := len(parts); l != 3 {
-		return "", "", commonerrors.NewCommandErrorMsgWithArgument(
-			commonerrors.ErrTypeMismatch,
-			fmt.Sprintf("Invalid payload: expected 3 parts, got %d", l),
-			"payload",
-		)
-	}
-
-	authzid, authcid, passwd := parts[0], parts[1], parts[2]
-
-	// Some drivers (Go) send empty authorization identity (authzid),
-	// while others (Java) set it to the same value as authentication identity (authcid)
-	// (see https://www.rfc-editor.org/rfc/rfc4616.html).
-	// Ignore authzid for now.
-	_ = authzid
-
-	return string(authcid), string(passwd), nil
-}
