+++ internal/handler/msg_saslstart_plain_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
+package handler
 
 import (
 	"encoding/base64"
